Fix default for enum options with a value
An enum in zserio can define a value for its options. If this is done
the first enum option may not have a value of 0, which results in no
option being marked as default. Instead of looking at a value of 0 we
no always mark the first option as default, which was always the
intended behavior.