Refactor decoding with type parameters
The generated rust code would fail when an array with a type paraemeter
using a function call was used. To fix this we now generate an initial
value with all non-index using parameters and clone that to all array
elements, and then loop over the array to fill in the missing values.